refactor: Внести финальные правки. Создать RowMapper для каждой модели.

+++ src/main/java/ru/yandex/practicum/filmorate/controller/GenreController.java
@@ -25,7 +25,7 @@ public class GenreController {
     }
 
     @GetMapping("/{id}")
-    public Genre getGenreById(@PathVariable int id) {
+    public Genre getGenreById(@PathVariable long id) {
         return filmService.getGenreById(id);
     }
